% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Data}
\alias{Data}
\title{An example of respiratory data}
\source{
Fantasia Database, PhysioNet
}
\description{
The respiratory data of a healthy people.
}
\details{
This is an data to be included in my package
}
\references{
"https://doi.org/10.13026/C2RG61"
}
\keyword{data}
