% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HqData.R
\docType{data}
\name{HqData}
\alias{HqData}
\title{The Hurst exponent of respiratory data}
\source{
Fantasia Database, PhysioNet
}
\description{
The Hurst exponent extracted from the MFDFA result of respiratory data.
}
\details{
This is an data to be included in my package
}
\references{
"https://doi.org/10.13026/C2RG61"
}
\keyword{data}
