% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MovingAverage.R
\name{MovingAverage}
\alias{MovingAverage}
\title{Function to calculate Moving Average of a series}
\usage{
MovingAverage(y, W)
}
\arguments{
\item{y}{a numeric vector, with respiratory data for a regularly spaced time series.}

\item{W}{a Positive integer, the windows of Moving Average.}
}
\value{
A new numeric vector after calculate the moving average.
}
\description{
function to calculate Moving Average of a series
}
\examples{
data("TestData")
W <- 50
y <- MovingAverage(Data[,2],W)
}
\references{
Zhang T, Dong X, Chen C, Wang D, Zhang XD. RespirAnalyzer: an R package for continuous monitoring of respiratory signals.
}
